Fixed LED light on pin 10

+++ IIDX 9+1e+9leds/Arduino Leonardo/leovx/leovx.ino
@@ -37,7 +37,7 @@ Encoder encL(0,1);
  * Turntable to pins 0 and 1
  */
 
-void lights(uint8_t lightDesc){
+void lights(uint16_t lightDesc){
   for(int i=0;i<buttonCount;i++){
      if((lightDesc>>i)&1){
          digitalWrite(ledPins[i],HIGH);
